+++ gatsby-node.js
@@ -154,6 +154,20 @@ exports.createPages = ({ graphql, actions }) => {
   })
 }
 
+// https://www.gatsbyjs.com/tutorial/authentication-tutorial/
+// Implement the Gatsby API “onCreatePage”. This is
+// called after every page is created.
+exports.onCreatePage = async ({ page, actions }) => {
+  const { createPage } = actions
+  // page.matchPath is a special key that's used for matching pages
+  // only on the client.
+  if (page.path.match(/^\/app/)) {
+    page.matchPath = "/app/*"
+    // Update the page.
+    createPage(page)
+  }
+}
+
 exports.onCreateNode = ({ node, actions, getNode }) => {
   const { createNodeField } = actions