+++ src/_reduxState/index.ts
@@ -7,6 +7,8 @@ import {
 } from "redux"
 import Cognito from "aws-sdk/clients/cognitoidentityserviceprovider"
 import { AWSError } from "aws-sdk"
+import { ApolloError } from "@apollo/client"
+import { isBrowser } from "utils"
 
 /**
  * action
@@ -50,7 +52,10 @@ type Data =
  * - saves the `data` to localstorage
  * - this could be scaled up by redux-persist
  */
-export const setCognito = (data: Data, error: AWSError | null): AnyAction => {
+export const setCognito = (
+  data: Data,
+  error: AWSError | ApolloError | null
+): AnyAction => {
   window.localStorage.setItem("cognito", JSON.stringify(data))
   return {
     type: ActionTypes.SET_COGNITO,
@@ -100,9 +105,7 @@ const initialState: RootState = {
   postsVersion: 1,
   [StateKeys.COGNITO]: {
     data: JSON.parse(
-      typeof window !== "undefined"
-        ? window.localStorage.getItem("cognito") ?? "{}"
-        : "{}"
+      isBrowser() ? window.localStorage.getItem("cognito") ?? "{}" : "{}"
     ),
     error: null,
     status: "idle",
@@ -135,7 +138,7 @@ const rootReducer: Reducer = (state = initialState, action) => {
 
 // Enable Redux Dev Tools
 const composeEnhancers =
-  typeof window === "object" && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
+  isBrowser() && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
     ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({
         // Specify extension’s options like name, actionsBlacklist, actionsCreators, serialize...
       })