+++ src/components/Form/Field.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { useField } from "formik"
+import { useField, ErrorMessage } from "formik"
 import styled, { BaseProps } from "styled-components"
 import { animated } from "react-spring"
 import theme from "styled-theming"
@@ -52,8 +52,8 @@ const FieldRenderer = styled(animated.div)`
     height: 2.8rem;
     background: ${background};
     border-color: ${theme("mode", {
-      light: props => (props.hasError ? "red" : borderColorBase),
-      dark: props => (props.hasError ? "darkred" : borderColorBase),
+      light: (props) => (props.hasError ? "red" : borderColorBase),
+      dark: (props) => (props.hasError ? "darkred" : borderColorBase),
     })};
     border-width: 1px;
     border-style: solid;
@@ -75,8 +75,8 @@ const FieldRenderer = styled(animated.div)`
 
   > input:focus {
     border-color: ${theme("mode", {
-      light: props => (props.hasError ? "red" : borderColorFocus),
-      dark: props => (props.hasError ? "darkred" : borderColorFocus),
+      light: (props) => (props.hasError ? "red" : borderColorFocus),
+      dark: (props) => (props.hasError ? "darkred" : borderColorFocus),
     })};
   }
 
@@ -106,8 +106,8 @@ const FieldRenderer = styled(animated.div)`
     text-transform: uppercase;
 
     color: ${theme("mode", {
-      light: props => (props.hasError ? "red" : color),
-      dark: props => (props.hasError ? "darkred" : color),
+      light: (props) => (props.hasError ? "red" : color),
+      dark: (props) => (props.hasError ? "darkred" : color),
     })};
   }
 `
@@ -141,9 +141,9 @@ export const Field = ({ label, ...props }: FieldProps) => {
     <FieldRenderer hasError={meta.touched && meta.error}>
       <input {...field} {...props} />
       <label htmlFor={props.id ?? props.name}>{label}</label>
-      {meta.touched && meta.error ? (
-        <FieldError>{meta.error}</FieldError>
-      ) : null}
+      <ErrorMessage name={props.name}>
+        {(msg) => <FieldError>{msg}</FieldError>}
+      </ErrorMessage>
     </FieldRenderer>
   )
 }